requirements.txt
cassandra-driver 
ipykernel
pandas
numpy
pathlib
scikit-learn
pyYAML


## Dev Requirements
pytest>=7.1.3
mypy>=0.971
flake8>=5.0.4
tox>=3.25.1
black>=22.8.0